Removed all remaining special treatment of the X500 common name, except one.
With network parameters the CN is no longer needed to identify notaries. This frees it up to be used in the node's name alongside the other attributes. The exception case is in devMode where the node CA cert is created with a subject name that is the node's configured name plus a CN value bolted on. IdentityService.verifyAndRegisterIdentity incorrectly relies on this to find the top-most identity cert. That needs fixing in a separate PR.

Also, the identity generation logic has been simplified, removing the need to have magic string values for storing distributed identities in the keystore. Now there are just two alias prefixes: "identity" as it was previously, and "distributed-notary".